+++ src/main/java/com/anthonyhilyard/advancementplaques/AdvancementPlaques.java
@@ -4,15 +4,12 @@ import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientLifecycleEvents;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvent;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import net.minecraftforge.api.ModLoadingContext;
+import net.minecraftforge.fml.config.ModConfig;
 
 
 public class AdvancementPlaques
 {
-	@SuppressWarnings("unused")
-	private static final Logger LOGGER = LogManager.getLogger();
-
 	public static final ResourceLocation TEXTURE_PLAQUES = new ResourceLocation(Loader.MODID, "textures/gui/plaques.png");
 	public static final ResourceLocation TEXTURE_PLAQUE_EFFECTS = new ResourceLocation(Loader.MODID, "textures/gui/plaqueeffect.png");
 
@@ -23,7 +20,7 @@ public class AdvancementPlaques
 
 	public static void onClientSetup()
 	{
-		AdvancementPlaquesConfig.init();
+		ModLoadingContext.registerConfig(Loader.MODID, ModConfig.Type.COMMON, AdvancementPlaquesConfig.SPEC);
 
 		ClientLifecycleEvents.CLIENT_STARTED.register((client) -> {
 			client.toast = new AdvancementPlaquesToastGui(client);