+++ src/main/java/com/anthonyhilyard/advancementplaques/AdvancementPlaquesConfig.java
@@ -1,76 +1,99 @@
 package com.anthonyhilyard.advancementplaques;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
-import me.shedaniel.autoconfig.AutoConfig;
-import me.shedaniel.autoconfig.ConfigData;
-import me.shedaniel.autoconfig.annotation.Config;
-import me.shedaniel.autoconfig.annotation.ConfigEntry;
-import me.shedaniel.autoconfig.serializer.JanksonConfigSerializer;
-import me.shedaniel.cloth.clothconfig.shadowed.blue.endless.jankson.Comment;
+import com.electronwill.nightconfig.core.Config;
 
+import org.apache.commons.lang3.tuple.Pair;
 
-@Config(name = "advancementplaques")
-public class AdvancementPlaquesConfig implements ConfigData
+import net.minecraft.resources.ResourceLocation;
+import net.minecraftforge.common.ForgeConfigSpec;
+import net.minecraftforge.common.ForgeConfigSpec.BooleanValue;
+import net.minecraftforge.common.ForgeConfigSpec.ConfigValue;
+import net.minecraftforge.common.ForgeConfigSpec.DoubleValue;
+import net.minecraftforge.common.ForgeConfigSpec.IntValue;
+import net.minecraftforge.common.ForgeConfigSpec.LongValue;
+
+public class AdvancementPlaquesConfig
 {
-	@ConfigEntry.Gui.Excluded
-	static AdvancementPlaquesConfig INSTANCE;
-
-	public static void init()
+	public static final ForgeConfigSpec SPEC;
+	public static final AdvancementPlaquesConfig INSTANCE;
+	static
 	{
-		AutoConfig.register(AdvancementPlaquesConfig.class, JanksonConfigSerializer::new);
-		INSTANCE = AutoConfig.getConfigHolder(AdvancementPlaquesConfig.class).getConfig();
+		Config.setInsertionOrderPreserved(true);
+		Pair<AdvancementPlaquesConfig, ForgeConfigSpec> specPair = new ForgeConfigSpec.Builder().configure(AdvancementPlaquesConfig::new);
+		SPEC = specPair.getRight();
+		INSTANCE = specPair.getLeft();
 	}
 
-	@Comment("If plaques should show on the top of the screen.")
-	public boolean onTop = true;
-	@Comment("The distance from the top or bottom of the screen, in pixels.")
-	public int distance = 16;
-	@Comment("The horizontal offset from the center, in pixels.")
-	public int horizontalOffset = 0;
-	@Comment("Hide waila/hwyla/jade popups while plaques are showing.")
-	public boolean hideWaila = false;
+	public final BooleanValue onTop;
+	public final IntValue distance;
+	public final IntValue horizontalOffset;
+	public final BooleanValue hideWaila;
 
-	@Comment("If plaques should show for task advancements (normal advancements).")
-	public boolean tasks = true;
-	@Comment("If plaques should show for goal advancements (medium-difficulty advancements).")
-	public boolean goals = true;
-	@Comment("If plaques should show for challenge advancements (high-difficulty advancements).")
-	public boolean challenges = true;
+	public final BooleanValue tasks;
+	public final BooleanValue goals;
+	public final BooleanValue challenges;
 
-	@Comment("Text color to use for plaque titles (like \"Advancement made!\"). Can be entered as an 8-digit LOWERCASE hex color code 0xaarrggbb for convenience.")
-	public long titleColor = 0xFF332200L;
-	@Comment("Text color to use for advancement names on plaques. Can be entered as an 8-digit LOWERCASE hex color code 0xaarrggbb for convenience.")
-	public long nameColor = 0xFFFFFFFFL;
+	public final LongValue titleColor;
+	public final LongValue nameColor;
+	
+	public final DoubleValue taskEffectFadeInTime;
+	public final DoubleValue taskEffectFadeOutTime;
+	public final DoubleValue taskDuration;
+	
+	public final DoubleValue goalEffectFadeInTime;
+	public final DoubleValue goalEffectFadeOutTime;
+	public final DoubleValue goalDuration;
 
-	@Comment("Duration of the shiny effect fade in for tasks.")
-	public double taskEffectFadeInTime = 0.5;
-	@Comment("Duration of the shiny effect fade out for tasks.")
-	public double taskEffectFadeOutTime = 1.5;
-	@Comment("Duration of the plaques for tasks (minus the effect fade in/out durations).")
-	public double taskDuration = 7.0;
+	public final DoubleValue challengeEffectFadeInTime;
+	public final DoubleValue challengeEffectFadeOutTime;
+	public final DoubleValue challengeDuration;
 
-	@Comment("Duration of the shiny effect fade in for goals.")
-	public double goalEffectFadeInTime = 0.5;
-	@Comment("Duration of the shiny effect fade out for goals.")
-	public double goalEffectFadeOutTime = 1.5;
-	@Comment("Duration of the plaques for goals (minus the effect fade in/out durations).")
-	public double goalDuration = 7.0;
+	public final ConfigValue<List<? extends String>> whitelist;
+	public final BooleanValue muteTasks;
+	public final BooleanValue muteGoals;
+	public final BooleanValue muteChallenges;
 
-	@Comment("Duration of the shiny effect fade in for challenges.")
-	public double challengeEffectFadeInTime = 1.25;
-	@Comment("Duration of the shiny effect fade out for challenges.")
-	public double challengeEffectFadeOutTime = 1.5;
-	@Comment("Duration of the plaques for challenges (minus the effect fade in/out durations).")
-	public double challengeDuration = 7.0;
+	public AdvancementPlaquesConfig(ForgeConfigSpec.Builder build)
+	{
+		build.comment("Client Configuration").push("client").push("visual_options");
 
-	@Comment("Whitelist of advancements to show plaques for.  Leave empty to display for all.")
-	public List<String> whitelist = new ArrayList<String>();
-	@Comment("If task sounds should be muted.")
-	public boolean muteTasks = false;
-	@Comment("If goal sounds should be muted.")
-	public boolean muteGoals = false;
-	@Comment("If challenge sounds should be muted.")
-	public boolean muteChallenges = false;
+		onTop = build.comment(" If plaques should show on the top of the screen.").define("on_top", true);
+		distance = build.comment(" The distance from the top or bottom of the screen, in pixels.").defineInRange("distance", 16, 8, 256);
+		horizontalOffset = build.comment(" The horizontal offset from the center, in pixels.").defineInRange("horizontal_offset", 0, -256, 256);
+		hideWaila = build.comment(" Hide waila/hwyla/jade popups while plaques are showing.").define("hide_waila", false);
+
+		tasks = build.comment(" If plaques should show for task advancements (normal advancements).").define("tasks", true);
+		goals = build.comment(" If plaques should show for goal advancements (medium-difficulty advancements).").define("goals", true);
+		challenges = build.comment(" If plaques should show for challenge advancements (high-difficulty advancements).").define("challenges", true);
+
+		titleColor = build.comment(" Text color to use for plaque titles (like \"Advancement made!\"). Can be entered as an 8-digit hex color code 0xAARRGGBB for convenience.").defineInRange("title_color", 0xFF332200L, 0x00000000L, 0xFFFFFFFFL);
+		nameColor = build.comment(" Text color to use for advancement names on plaques. Can be entered as an 8-digit hex color code 0xAARRGGBB for convenience.").defineInRange("name_color", 0xFFFFFFFFL, 0x00000000L, 0xFFFFFFFFL);
+
+		build.pop().push("duration_options");
+
+		taskEffectFadeInTime = build.comment(" Duration of the shiny effect fade in for tasks.").defineInRange("task_effect_fadein", 0.5, 0.1, 3.0);
+		taskEffectFadeOutTime = build.comment(" Duration of the shiny effect fade out for tasks.").defineInRange("task_effect_fadeout", 1.5, 0.1, 3.0);
+		taskDuration = build.comment(" Duration of the plaques for tasks (minus the effect fade in/out durations).").defineInRange("task_duration", 7.0, 2.0, 10.0);
+		
+		goalEffectFadeInTime = build.comment(" Duration of the shiny effect fade in for goals.").defineInRange("goal_effect_fadein", 0.5, 0.1, 3.0);
+		goalEffectFadeOutTime = build.comment(" Duration of the shiny effect fade out for goals.").defineInRange("goal_effect_fadeout", 1.5, 0.1, 3.0);
+		goalDuration = build.comment(" Duration of the plaques for goals (minus the effect fade in/out durations).").defineInRange("goal_duration", 7.0, 2.0, 10.0);
+
+		challengeEffectFadeInTime = build.comment(" Duration of the shiny effect fade in for challenges.").defineInRange("challenge_effect_fadein", 1.25, 0.1, 3.0);
+		challengeEffectFadeOutTime = build.comment(" Duration of the shiny effect fade out for challenges.").defineInRange("challenge_effect_fadeout", 1.5, 0.1, 3.0);
+		challengeDuration = build.comment(" Duration of the plaques for challenges (minus the effect fade in/out durations).").defineInRange("challenge_duration", 7.0, 2.0, 10.0);
+
+		build.pop().push("functionality_options");
+
+		whitelist = build.comment(" Whitelist of advancements to show plaques for.  Leave empty to display for all.").defineListAllowEmpty(Arrays.asList("whitelist"), () -> new ArrayList<String>(), e -> ResourceLocation.isValidResourceLocation((String)e) );
+		muteTasks = build.comment(" If task sounds should be muted.").define("mute_tasks", false);
+		muteGoals = build.comment(" If goal sounds should be muted.").define("mute_goals", false);
+		muteChallenges = build.comment(" If challenge sounds should be muted.").define("mute_challenges", false);
+
+		build.pop().pop();
+	}
 }